+++ Editor/Source/Events.cs
@@ -1,11 +1,10 @@
-﻿namespace GameDBEditorLibrary {
+﻿namespace GameDBEditorLibrary
+{
     internal enum Events
     {
         GAMEDB_LOADED,
         RELOAD_TABLE,
         EDIT_TABLE,
-        REVISION_LOADED,
-        REVISION_UNLOADED,
         DELETE_TABLE
     }
 }