+++ Runtime/Core/Server/RequestMethod.cs
@@ -1,5 +1,6 @@
 ﻿namespace GameDBLibrary
 {
+    [System.Obsolete(LegacyRemoteApi.Message)]
     public
         enum RequestMethod
     {